Refactor : custom Object 폴더 생성 및 분리
