+++ backend/src/album/album.entity.ts
@@ -1,5 +1,5 @@
 import { Group } from "src/group/group.entity";
-import { TimeStampEntity } from "src/myBaseEntity/timestampEntity";
+import { TimeStampEntity } from "src/custom/myBaseEntity/timestampEntity";
 import { Post } from "src/post/post.entity";
 import { Column, Entity, JoinColumn, ManyToOne, OneToMany, PrimaryGeneratedColumn } from "typeorm";
